Convert to separate package
Added __init__.py.
Breaking Process class into two separate files
