+++ sudio/sudio.py
@@ -1,0 +1,13 @@
+# the name of allah
+from . import _process
+from ._register import Members as Mem
+
+@Mem.sudio.parent
+class Sudio:
+    class Pipeline:
+        def __init__(self):
+            pass
+
+        def new_element(self, callback):
+            pass
+
